GPU-PHY S-7f-2b: hold the cached plan's address in a pointer, not an unsigned
The hop-table cache never hit: its key was the plan's address stored in an `unsigned`, so on a 64-bit
host the comparison against `&plan` could never match and every dispatch rebuilt and re-uploaded the
tables - exactly the cost the previous commit set out to remove. The counter said 4 builds and 0
reuses per hop (one per dispatch), and with the key fixed the same hop is 1 build and 3 reuses.

Gates: 980 recorded receptions replay to byte-identical LLRs on the device and the host route,
ctest -L phy 162/162, the equalizer unit test is green, and 120 consecutive replays run without a
crash.

+++ lib/phy/upper/channel_processors/metal/ocudu_equalizer_metal_engine.mm
@@ -555,9 +555,11 @@ struct eq_flush_state_t {
   std::vector<id<MTLBuffer>> gather_tables;
   /// The hop's tables while the burst is being encoded (see eq_gather_tables): one build per hop,
   /// every dispatch of that hop reads them.
-  gather_tables_t  hop_tables;
-  unsigned         hop_tables_plan = 0; // the plan they were built from, as an address
-  bool             hop_tables_valid = false;
+  gather_tables_t hop_tables;
+  /// The plan they were built from, as an address. It must hold a POINTER: an unsigned truncated it
+  /// on a 64-bit host and the key never matched the plan it was made from, so the cache never hit.
+  uintptr_t hop_tables_plan  = 0;
+  bool      hop_tables_valid = false;
 };
 static eq_flush_state_t& eq_flush_state()
 {
